+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/fishlo/v1/fishlo/program/Competition/AutoWobbleRight.java
@@ -5,7 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import org.firstinspires.ftc.teamcode.fishlo.v1.fishlo.program.FishloAutonomousProgram;
 import org.firstinspires.ftc.teamcode.fishlo.v1.fishlo.robot.OpenCV;
-import org.firstinspires.ftc.teamcode.fishlo.v1.fishlo.robot.Utility.PID;
+import org.firstinspires.ftc.teamcode.fishlo.v1.fishlo.robot.Util.PID;
 import org.firstinspires.ftc.teamcode.robot.Robot;
 
 @Autonomous
